Move NetworkProvider enum into its own file
