+++ src/config/network/environments/local.ts
@@ -1,4 +1,4 @@
-import { NetworkProvider } from '@/util/network';
+import { NetworkProvider } from '@/types/enums';
 
 export const local = {
     [NetworkProvider.Test]: [
