+++ src/config/network/environments/prod.ts
@@ -1,4 +1,4 @@
-import { NetworkProvider } from '@/util/network';
+import { NetworkProvider } from '@/types/enums';
 
 export const prod = {
     [NetworkProvider.Test]: [
