+++ src/controllers/asset_pools/assetPools.test.ts
@@ -1,6 +1,6 @@
 import request from 'supertest';
 import app from '@/app';
-import { NetworkProvider } from '@/util/network';
+import { NetworkProvider } from '@/types/enums';
 import { timeTravel, signMethod, deployExampleToken, createWallet } from '@/util/jest/network';
 import {
     rewardWithdrawAmount,
