+++ src/controllers/deposits/deposits.test.ts
@@ -6,7 +6,8 @@ import { Contract } from 'web3-eth-contract';
 import { afterAllCallback, beforeAllCallback } from '@/util/jest/config';
 import { getToken } from '@/util/jest/jwt';
 import { agenda, eventNameRequireDeposits } from '@/util/agenda';
-import { NetworkProvider, solutionContract } from '@/util/network';
+import { solutionContract } from '@/util/network';
+import { NetworkProvider } from '@/types/enums';
 import { IPromoCodeResponse } from '@/types/interfaces/IPromoCodeResponse';
 import { createWallet, deployExampleToken } from '@/util/jest/network';
 import { findEvent, parseLogs } from '@/util/events';
