+++ src/controllers/members/members.test.ts
@@ -1,6 +1,7 @@
 import request from 'supertest';
 import app from '@/app';
-import { getProvider, NetworkProvider } from '@/util/network';
+import { getProvider } from '@/util/network';
+import { NetworkProvider } from '@/types/enums';
 import { createWallet, deployExampleToken, voter } from '@/util/jest/network';
 import { userWalletPrivateKey2 } from '@/util/jest/constants';
 import { Contract } from 'web3-eth-contract';
