+++ src/controllers/promo_codes/promoCodes.test.ts
@@ -5,7 +5,7 @@ import { Contract } from 'web3-eth-contract';
 import { getToken } from '@/util/jest/jwt';
 import { IPromoCodeResponse } from '@/types/interfaces/IPromoCodeResponse';
 import { afterAllCallback, beforeAllCallback } from '@/util/jest/config';
-import { NetworkProvider } from '@/util/network';
+import { NetworkProvider } from '@/types/enums';
 import { deployExampleToken } from '@/util/jest/network';
 
 const http = request.agent(app);
