+++ src/services/WithdrawalService.ts
@@ -1,5 +1,6 @@
 import { toWei, fromWei } from 'web3-utils';
-import { NetworkProvider, MaxFeePerGasExceededError } from '@/util/network';
+import { MaxFeePerGasExceededError } from '@/util/network';
+import { NetworkProvider } from '@/types/enums';
 import { WithdrawalState, WithdrawalType } from '@/types/enums';
 import { AssetPoolType } from '@/models/AssetPool';
 import { Withdrawal } from '@/models/Withdrawal';
