+++ src/types/enums/NetworkProvider.ts
@@ -1,0 +1,4 @@
+export enum NetworkProvider {
+    Test = 0,
+    Main = 1,
+}
