+++ src/types/enums/index.ts
@@ -2,3 +2,4 @@ export * from './WithdrawalState';
 export * from './WithdrawalType';
 export * from './DepositState';
 export * from './GasAdminType';
+export * from './NetworkProvider';
