+++ src/util/jest/network.ts
@@ -2,7 +2,8 @@ import { Account } from 'web3-core';
 import { Artifacts } from '@/util/artifacts';
 import { soliditySha3 } from 'web3-utils';
 import { VOTER_PK, DEPOSITOR_PK, mintAmount } from './constants';
-import { getProvider, NetworkProvider, solutionContract } from '@/util/network';
+import { getProvider, solutionContract } from '@/util/network';
+import { NetworkProvider } from '@/types/enums';
 import { TransactionService } from '@/services/TransactionService';
 
 const { web3, admin } = getProvider(NetworkProvider.Main);
